Add files via upload
The Arduino_Serial_PWM..ino is a code that can receive data from atoi object. The Arduino_Eurorack_Quad_DAC_Volatage_Ramp.ino is an modified example from Adafruit library used to test voltage ramp. So for the next step, I want to combine these two process, I want to use max as the main control to send data to Arduino and the quad DAC, so the voltage can be controlled by Max.